listener manager: implement dynamic add/remove (#1215)
This is the main heavy lifting PR to add dynamic add/removal of
listeners at runtime. There are various things still to do (see
TODOs in code) which will be added in follow up changes. The
actual v1 LDS API using RestApiFetcher will also be added in a follow
up.

This commit contains both unit tests as well as a full end to end
integration test that demonstrates adding and removing a listener
in the running server. In the future we should extend this
integration test to also cover dynamic cluster add/remove.