proxy_protocol: Use actual local address.
If the listening socket is bound to an all nodes IP address (e.g.,
0.0.0.0), the connection object will have that as it's local address,
too.  This is conceptually wrong, as the connection's actual
destination address is never the all nodes IP address, but the actual
destination IP address the client made the connection with.  Fix this
for the proxy protocol implementation and add some basic tests.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>