proxy_protocol: Use both addresses, ports.
So far only the remote address parsed from the proxy protocol v1
header was used when creating a new connection.  Use both addresses as
well as ports, so that any filters that may refer to the addresses see
them.

This also gets around of the all hosts address of a wildcard listener
passing through as the destination address of a connection.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>