lds: implement the the v1 API (#1238)
This is a v1 (in the API sense) implementation of LDS. The work to get
the v2 API running is substantial and I don't want to block initial users
from testing out the API since the plumbing is the hard part that won't
change between API versions. There are still some follow up TODOs but
this should be fully functional for beta use.