cluster manager: fix crash related to "local cluster" handling
This commit fixes a crashing issue in which a CDS cluster is removed,
leaving behind a stale member update callback inside the "local cluster."
The fix is to allow member update callbacks to be removed. However, as
part of investigating this issue I uncovered several related issues:
1) We will leak TLS slots in several places when using LDS.
2) Redis is broken in two ways regarding LDS/CDS. This commit prevents
   one of them from hapening by just prohibiting Redis from working with
   a CDS cluster (for now), but LDS is still broken in a similar way WRT
   the member update callbacks.

Fixing both of the above issues is not trivial so I will do them in follow
ups since this change is large enough as it is. Also, fixing the Redis LDS
case will make a portion of this fix cleaner.

Fixes https://github.com/lyft/envoy/issues/1254