load_balancer: Compute hash on-demand. (#1301)
The load balancer context is specified as a pure virtual class, but
the return type of the hashKey() member implies that there is a value
stored inside the implementation, since a const reference is returned.
Change the return type to a non-const non-reference, allowing a
stateless implementation.  Enhance the one call site that called
hashKey() twice to call it only once to avoid overhead in case each
hashKey() call recomputes the hash.

Change the router Filter class to inherit from the LoadBalancerContext
interface, allowing for simpler, on-demand implementation of the load
balancer context.

With this change a load balancer context never needs to be dynamically
allocated.  This will make exteonsions to load balancer context also
simpler and cheaper.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>