tls: refactor tls to work better with lds and cds
Previously, TLS slots would be leaked if either clusters or listeners
transitively allocated a slot. This commit refactors TLS so that slots
are hidden behind an interface and are freed properly. As part of this
change shutdown() has been removed from the thread local object interface
so that proper RAII semantics can be used. The one special case to this
relates to stat flushing; special code is needed so that once shutdown
begins no flushing of timers/histograms happens. It seems better to
have this one special case vs. requiring everyone to implement a
shutdown() method.

A few test changes are also included:
1) We now use the real thread local stat store in integration tests since
   it provides better coverage.
2) Added two integration tests that cover the server being shutdown with
   active connection pool connections which expose TLS ordering issues.