http: convert HTTP/2 setting JSON handling to protobuf.
As a step towards CDS v2 support, this PR introduces a translation library, ProtocolJson, which
translates v1 JSON for HTTP/2 settings to the corresponding v2 Http2ProtocolOptions message,
alongside some support utility classes to assist with translating to/from JSON/proto. This is the
stylized approached (translateSomeMessage(...)) that we will be using to transform the v1 JSON
objects into v2 protos.

The PR is effectively a nop, since it translate from JSON to proto and then consumes from proto
with the same semantics as the JSON accessors previously. There is a small change however to HTTP/2
settings processing; we only consider the no_compression HTTP option, since it's the only one
allowed by the schema.

As a bonus, there are a handful of additional helpers for Duration added that will be used in later PRs.
Since they are largely similar to their UInt32 counterparts, it made sense to add them here.