http: add setStatic() header string function
I was looking at coverage and I saw that the new http/2 static header
code for values was not covered. This commit adds a new setStatic()
function on header string and uses it in a few high throughput cases.
This gets coverage and also should be a bit better for perf.

Also fix two random missing coverage lines.