http: add setReference() header string function (#1360)
I was looking at coverage and I saw that the new http/2 static header
code for values was not covered. This commit adds a new setReference()
function on header string and uses it in a few high throughput cases.
This gets coverage and also should be a bit better for perf.

Also do some renames for clarity.

Also fix two random missing coverage lines.