introduce singleton manager and fix multiple TLS date providers
It is a common problem that there needs to be a single instance of
something in the server. This change introduces the concept of the
singleton manager which can be used for that purpose. All singletons
are looked up by name and must be statically registered like existing
filters. This avoids collision.

This change uses the singleton manager to create a single TLS date provider
for the entire server.

Fixes https://github.com/lyft/envoy/issues/1362.