bazel: non-git version info for distributions
When packaging envoy for a distribution package (rpm, deb, etc), the
source will often not be a `git clone` but a downloaded zip or tar
archive from the project.

This avoids the use of calls to `git` for compiling in version
information to the binary, and instead allows for distributions to
provide a SOURCE_VERSION file at the root of the source tree.
If present, that information will be used instead. If the file isn't
present, continue as usual to calling `git`.

For testing, I `echo -n %{git_commit} > SOURCE_VERSION`, and got the
following output:
```
$ bazel --version

bazel version: 5ad63908951d98a5bc31352d8ac218b6590d4b12/Distribution/DEBUG
```

Signed-off-by: Vincent Batts <vbatts@hashbangbash.com>