router: hard drop support for virtual cluster priorities. (#1412)
We removed these from the v2 API, since priorities are not a stats concept (which VCs are intended
to be a logical grouping for) and instead affect the route action. We're not aware of any use of
this feature today, this PR attempts to simplify the v2 RDS protoization by doing a hard drop of
support (rather than following the usual deprecation process).