Clean up test/coverage/BUILD after coverage run. (#1486)
This file is autogenerated. If you do a coverage build, then change
other BUILD files to remove something referenced in the autogenerated
file, Bazel breaks in a confusing manner that is NOT resolved by "bazel
clean" and the presence of this file isn't obvious due to .gitignore.

It can be safely removed after the coverage build executes.