config: support for YAML parsing.
This patch introduces support for YAML parsing in Envoy for the v1 and v2 APIs. If an input file
path ends with ".yaml", it will be parsed with yaml-cpp. An internal Json::Object is constructed
with the parsed YAML::Node.

There is one quirk that exists due to the schemaless parsing used here (and desirable for v2 due to
the opaque filter configs). YAML relies on the application or schema to determine what happens when
there is ambiguity in a scalar value between bool/integer/float/string. In the Envoy parsing, we
rely on the user disambiguating fields to support a schemaless parse. E.g. to avoid a cluster name
1234 being parsed as int, it will have to be written as "1234". Same goes for using 1.0 for a float
field to avoid treatment as an integer if only 1 is provided.