outlier: fix consecutive 5xx without interleaved success (#1493)
Fixes #1491.

The DetectorHostSink has a consecutive_5xx_ counter that gets incremented every time a 5xx
response gets charged to the sink, and reset every time a non-5xx does. Previously if a Sink 
reached the consecutive 5xx limit, and the host was charged with a consecutive 5xx ejection, 
the consecutive_5xx_ counter was not been reset. This meant that the Sink would not trigger
another consecutive 5xx ejection until a non-5xx code was charged and the counter was reset,
and then subsequently increased to the limit. This PR fixes the issue by resetting the counter 
after the host has been charged with an ejection (whether the ejection is enforced or not).