Support bootstrap configs in text proto format, and improve errors.
This commit allows `.pb_text` configs to be parsed as text, and improves
the parse failure error message to include (1) text or binary mode, and
(2) the message type being parsed.

I was trying to use the new `--bootstrap-path` flag to get away from
JSON so the config could have comments, and got stuck for a while before
realizing the protobuf was being parsed in _binary_ mode.