config: ADS plumbing. (#1602)
This patch introduces AdsApi, which is where the ADS implementation will go, and AdsSubscription,
that matches between the typed Subscription model and untyped AdsApi interface. AdsApi is placed in
ClusterManager, despite being only dubiously related to ClusterManager, since it is needed early on
and is codependent on ClusterManager. It's also needed at all the xDS instantiation sites where
ClusterManager is.