Add static registration of AccessLog::Instance implementations. (#1609)
This enables custom access log implementations to be substituted in place of the existing file access log. This feature is only available in the v2 api. For v1 api users, there should be no changes. This change increments the commit hash of envoy-api to get related api changes.

The implementation follows the static registration pattern for other components relatively closely.