Switch to return-by-value for versionInfo string.
Returning the versionInfo by reference creates problems for the Google
import due to the protobuf generated code returning a string variant
convertible to std::string rather than std::string itself. This results
in the creation of a temporary std::string in grpc_subscription_impl.h
and http_subscription_impl.h, which cannot be returned by reference.

Switch to return-by-value for the versionInfo string.