server: fix SIGTERM during init crashes
Fixes https://github.com/lyft/envoy/issues/1610

If we receive SIGTERM during initialize we can attempt to start
workers while shutting down. This fixes that. A small amount of
code was broken out to ease testing. I'm not crazy about that, but
I could not find a better way to test this case without a bunch
more effort which did not seem worth it.