config: latch previous config on empty EDS/RDS update. (#1688)
With ADS or other batched updates, we might not see a cluster/route config in
an EDS/RDS update. There are at least two sane things to do; (1) latch the
previous config (2) provide an empty config for the resource. In this PR, we
opt for (1), this isn't on the happy path so there isn't an obvious choice
either way.

Signed-off-by: Harvey Tuch <htuch@google.com>