Use originating client IP address for hash routing
Bugfix: the previous version of this code used the address of the
directly connected downstream host. This is not the same as the address
of the client making the request unless the client is directly connected
to the Envoy instance doing the routing. To fix this, use the downstream
address of the routed stream instead.

Signed-off-by: Alex Konradi <akonradi@google.com>