CircleCI basic support (#1716)
This commit gives us basic IPv4 support via CircleCI. We still test IPv6 on
Travis as Circle does not yet support IPv4. This commit is going to fail the
master pushes. I need to make new keys and creds and configure env vars
in Circle. I would like to ship this and then I will work on that.

Signed-off-by: Matt Klein <mklein@lyft.com>