router: add x-envoy-overloaded header
This change reduces retry explosion:
1) Retry is never performed if local circuit breaking occurred.
2) If traffic was dropped (circuit breaking or maintenance mode),
   Envoy will set an x-envoy-overloaded header. (Right now it is
   set to "true" which has no real meaning. We might define more
   values in the future).
3) Envoy will not retry if it sees the x-envoy-overloaded header
   set by upstream. Of course this header can be propagated
   downstream by the caller if desired and application other than
   Envoy can set this header. However, even in the Envoy <-> Envoy
   single hop case this can reduce retry explosion when under
   duress.

One downside of this change is that a single bad upstream might
create a situation in which retries do not happen that might
otherwise succeed. In general, it seems better to err on the side
of less retries and bad single hosts should be ejected via proper
outlier detection policies.

Signed-off-by: Matt Klein <mklein@lyft.com>