Update stats interface in preparation for tags (#1803)
Tags will require all stats have non-trivial state associated with them. This enforces that all stats are represented as objects before being exported. As a part of this addition, there are a few structural changes to the way stats work in Envoy:

* The Metric calss has been added as a base class for all stats objects. This will contain identifying information.
* The Histogram class has replaced the Timer class. This object represents histograms of any data that can be represented by uint64_t. It also contains a method type() to identify the type of data represented (this includes Duration).
* Timespan has been moved into an independent class, which will be created with a reference Histogram.
* HISTOGRAM macros have been added to create Integer histogram objects.
This is the first stage of what was previously #1751 with some modifications based on the discussion in that PR.

Signed-off-by: Matt Rice <mattrice@google.com>