upstream: switch dropped stats accounting from host to cluster.
It makes no sense to attribute dropped stats to the host (or locality), since the drop decision is
made at the cluster level, prior to host picking. Instead, introduced a parallel concept of cluster
stats, load report stats, that don't get sent to sinks and can be latched independently of the stats
flush period.

As a bonus, bumped data-plane-api SHA to bring in corresponding EDS load report proto changes.

Signed-off-by: Harvey Tuch <htuch@google.com>