load balancer: thread metadata matches through LoadBalancerContext (#1844)
3rd of 4 reviews split out from #1735.

Router::MetadataMatchCriteria is implemented. Instances of the implementation are created when a route's RouteAction has an "envoy.lb" section in metadata_match values. These are merged with additional "envoy.lb" metadata_match values from the WeightedCluster, if any. In the event of a collision, the WeightedCluster keys take precedence.

The LoadBalancerContext interface is modified to allow MetadataMatchCriteria to be accessed by load balancers. Router::Filter's implementation of the LoadBalancerContext is updated to return the selected route's criteria, if any.

Signed-off-by: Stephan Zuercher <stephan@turbinelabs.io>