Fix warning about integer overflow in `WatcherImpl::onInotifyEvent()`.
This change maintains the current behavior by skipping the event loop if
`rc < 0`, and fixes the warning by converting `index` to unsigned `size_t`.

Changing the type of `buffer` to `uint8_t[]` fixes another warning about
not NUL-terminating strings. Coverity thinks a `char[]` should always
be NUL-terminated, which is reasonable, but here we're only using it
as a plain memory buffer.

Signed-off-by: John Millikin <jmillikin@stripe.com>