http: close conn pool connections on HC failure
This change implements new functionality which will close HTTP
conn pool connections when a host shifts to a health check failed
status (either active or passive). The idea behind this change
is that a connection may become bad (potentially due to specific
ECMP route), and especially for HTTP/2, Envoy may continue to
use a bad connection for a very long time if a FIN is never
received. This change will maximize the chance that when a host
gets pulled back into rotation with fresh connections the
performance will be better.

We will likely add similar functionality for Redis in the future.

Signed-off-by: Matt Klein <mklein@lyft.com>