mongo: add runtime feature flag for drain close behavior
There are some issues in Lyft's environment related to this feature.
I need to do further work on draining in general but that is going
to take me several days. This is a short term fix and seems fine to
have this control here in either case.

Note that I will add better documentation on draining as an overall
topic when I do the draining follow ups.

Signed-off-by: Matt Klein <mklein@lyft.com>