proxy protocol: parse UNKNOWN protocol (#1728)
Ignore any parameters sent on the PROXY line when the procotol is
UNKNOWN and accept the connection with current settings.

As per protocol spec:

If the announced transport protocol is "UNKNOWN", then the receiver
knows that the sender speaks the correct PROXY protocol with
the appropriate version, and SHOULD accept the connection and use
the real connection's parameters as if there were no PROXY protocol
header on the wire.

Signed-off-by: Filip Andres <filip@andresovi.net>