server: Fix assert failure during early shutdown
Currently we assert that the workers have been started before stopping
them, but we also avoid starting the workers if we've already received
SIGTERM by that point in initialization. The result is that sending
Envoy a SIGTERM too early leads to an assert failure (in builds with
asserts enabled).

This changes the assert to an early return, so that we simply skip the
calls to both start() and stop() if shutting down early, and adds a test
to cover that case.

Signed-off-by: Reuven Lazarus <rzl@google.com>