config: --v2-config-only CLI flag.
This determines whether the config should only be parsed as v2. If false, when a v2 parse fails,
a second attempt to parse the config as v1 will be made (existing behavior).

This improves the UX when working with v2 configs, as Envoy will fail hard with meaningful messages
rather than attempting to fallback and presenting v1 config errors when a bad v2 config parses as
neither v1 or v2.

When we deprecate v1, this flag will default to true.

Signed-off-by: Harvey Tuch <htuch@google.com>