config: v1 routes must include redirect or cluster
This is a bugfix to disallow 'bridge to nowhere' v1 route
configurations, as described by the docs.[1] This prevents an outage
case where operator error results in a route definition which matches
traffic but does not route it.

The v2 API is not vulnerable to this error because of the Protobuf oneof
constraint applied to (RouteAction|RedirectAction), the RouteAction
cluster_specifier constraint, and the required
host_redirect/path_redirect fields in RedirectAction.

Risk level: Low
This is a breaking change for v1 route configs, but I don't think there
are any valid reasons to have a bridge to nowhere route config.

Testing: unit test

[1] "If the route is not a redirect (host_redirect and/or
path_redirect is not specified), one of cluster, cluster_header, or
weighted_clusters must be specified."

Signed-off-by: Ben Tyler <benjamin.tyler@booking.com>