server: If init fails, return 1 from main_common instead of calling exit(1).
This change takes the exit(1) out of the Server::InstanceImpl
constructor, and replaces it with a return 1 from main_common by way of
an EnvoyException from the constructor. The reason is that at Google we
have some cleanup work to do after main_common returns (whether success
or failure). If it calls exit() instead of returning, we never get the
chance.

Signed-off-by: Reuven Lazarus <rzl@google.com>