server: Throw exceptions from OptionsImpl ctor instead of calling exit.
This is a logical continuation to #2157. The OptionsImpl constructor
used to call exit(0) when the argv was e.g. "envoy --hot-restart-version"
(which prints a message and exits) or exit(1) when it was malformed.
With this change, it throws exceptions in those cases instead.

The exceptions are caught in main(), which returns 0 or 1 accordingly.
Envoy users who write their own main() will want to do the same when
they construct OptionsImpl (after doing any necessary cleanup work), so
this is a more significant developer-facing API change than #2157 was.

Signed-off-by: Reuven Lazarus <rzl@google.com>