server: Throw exceptions from OptionsImpl ctor instead of calling exit. (#2177)
This is a logical continuation to #2157. The OptionsImpl constructor used to call exit(0) when the argv was e.g. "envoy --hot-restart-version" (which prints a message and exits) or exit(1) when it was malformed. With this change, the constructor throws exceptions in those cases instead.

The exceptions are caught in main(), which returns 0 or 1 accordingly. Envoy users who write their own main() will want to do the same when they construct OptionsImpl (after doing any necessary cleanup work), so this is a more significant developer-facing API change than #2157 was.

Risk Level: Medium

Testing:
Updated unit tests to verify the exception is thrown. As a nice side benefit, we don't need death tests for those cases anymore.

Docs Changes:
N/A

Release Notes:
N/A

Signed-off-by: Reuven Lazarus <rzl@google.com>