access log: consolidate and clean up address handling
This is a wide reaching change that consolidates address handling
in request info / access logs to use real addresses vs. strings.
This will make it substantially easier to fill in protos when we
send gRPC access logs. It also has the added benefit of making
all address handling well typed. As part of this change I also
did the following:

- Added a new DOWNSTREAM_LOCAL_ADDRESS formatter.
- Made DOWNSTREAM_ADDRESS/CLIENT_IP print only the IP and not
  the port for IP addresses. This was done because of back-compat
  reasons for HTTP access logs. This will be a behavior change for
  TCP access logs. I don't see any way around this unfortunately
  without doing a ton of extra ugly work. Given that the % of users
  of HTTP logs dwarfs the % of users of TCP logs and that generally
  users do not care about the port of the remote, I think this is OK.
  If people disagree we can reconsider.
- Was able to cleanup various string mutations and parsing in the
  connection manager utility path as we are now operating on
  real addresses in many cases.

Signed-off-by: Matt Klein <mklein@lyft.com>