http conn man: fix x-envoy-internal regression
This fixes a regression from https://github.com/envoyproxy/envoy/pull/2217.
In that PR, we refactored calculation of downstream address to use real
addresses and not strings. As part of that change, I broke some behavior
that people were depending on. Primarily that x-envoy-internal is based on:
1) XFF is present (either via use_remote_address or in request headers).
2) XFF contains a single address after any appending from use_remote_address.
3) The address is valid.

In all other cases the request is not considered internal. The above algorithm
is IMO not optimal in many ways, but it's how the code has worked for a long
time and Lyft (and possibly) others are depending on it, so we can't change it.

This change restores the old behavior. It reverts changes to integration tests
for checking for bytes, as well as does a substantial cleanup to the connection
manager utility tests since they were a mess. It also adds some new tests to
specifically cover certain cases.

In the future I think we will want to add additional inference modes for
determining internal requests, but for now, we need to keep the old behavior.

Signed-off-by: Matt Klein <mklein@lyft.com>