Add an HTML home page for the admin port, which makes it easier to surf to the other handlers with a browser.
Add a private mechanism to have admin handlers supply arbitrary
headers.  Note that the handler callback doesn't expose a header
structure, so the public v2 API does not provide this header-control.
However, it was specifying text/html as content-type by default, so
I changed the default to text/plain (with nosniff).

Added sanitization of both the prefix string and the help text, which
need to be safely injected into the HTML doc.

Added some styling and a reference to the existing Envoy favicon as well.

Note that no existing handlers get any response bytes changed as a result
of this commit, just their content-type.

Signed-off-by: Joshua Marantz <jmarantz@google.com>