Add an HTML home page for the admin port, which makes it easier to surf to the other handlers with a browser.
Make the admin HandlerCb pass through a headers reference so handlers
can optionally set the content-type and/or cache-control, changing the
defaults to text/plain and no-cache, respectively.

Added sanitization of both the prefix string and the help text, which
need to be safely injected into the HTML doc.

Added some styling and a reference to the existing Envoy favicon as
well.

Added some tests and several test-variants to existing tests.

Signed-off-by: Joshua Marantz <jmarantz@google.com>