Automatically add DCO signoff with commit hooks (#2283)
Fixes #2281. DCO signoff is mandatory for all contributors. Currently
this is enforced via CI.

This commit smoothes this process somewhat by introducing:

A git commit hook, which will check the commit message for a
signoff, and if not present, add one to the bottom of the message.
A script, support/bootstrap, which will symlink these commit
hooks into the Envoy root .git directory, to minimize the effort
it takes to use them.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>