stats: make prometheus formatter unit-testable
There is a set of unit tests that can verify the sanitization of metrics
names when exported to Prometheus.  These rely on the default set of
metrics for a fresh server which, incidentally, doesn't include metrics
with some prohibited characters. #2263 reports one of these cases, '-'.
An integration test falls in a similar case.

A fix for #2263 is trivial, but adding a regression test is not.  The
root reason is that the affected method is private in AdminImpl, so the
only way of hitting the right code path is to add excessive complexity
as both unit/integration versions would require configuration that is
known to generate metrics with '-', etc.

This commit makes a preliminary refactor on admin.cc, moving all code
involved in formatting metrics for Prometheus to its own class where
methods can be made public (instead of polluting AdminImpl's API.) A
later improvement could also refactor the JSON formatter to a separate
class and introduce a strategy pattern for formatters.  This requires a
shuffling a bit more code so I preferred to take a single step here.

Tests are supplied, matching the current behaviour, but still incorrect
as per #2263.  A later commit will fix #2263, and update the test
accordingly.

Signed-off-by: Galo Navarro <anglorvaroa@gmail.com>