stats: prometheus metrics sanitize '-' fixes #2263
The Prometheus metrics formatter only sanitizes '.', but leaves other
non supported chars like '-'.  A case where this may appear is on cipher
suites as reported in #2263.

The change simply adds a second replace for '-' (rather than a regex to
avoid the perf hit, same reason for not matching on all [^a-Z0-9_]
chars.)

This replacement is only applied to metric and tag names, not to tag
values.

Signed-off-by: Galo Navarro <anglorvaroa@gmail.com>