http conn pool: rework closeConnections() -> drainConnections()
This is a follow up from https://github.com/envoyproxy/envoy/pull/2005.
From production usage at Lyft it has become clear that immediately
closing connections on health check failure is too aggressive. During
normal upstream draining when using the x-envoy-immediate-health-check-fail
header, it is possible to close connections while they still have legitimate
draining requests.

This change reworks the functionality to drain but not close existing
connections. This will make it so that *new* streams occur on new connections,
but existing streams are allowed to finish on their existing connections.
This is the behavior we actually wanted in the first place.

Signed-off-by: Matt Klein <mklein@lyft.com>