Avoid having crawlers mutate server state.
Keeps track of whether admin handlers mutate state or not.  When
showing the root page, use a form with a button, rather than an a-tag
to reference URLs for mutating state.

Also: in preparing the sorted-list of handlers, use a std::vector and
std::sort rather than using a map to get the sorting behavior, as it's
generally faster, since (a) it doesn't have to keep the collection
sorted as its being incrementally inserted and (b) it doesn't have to
create as many internal structures with pointers.  This doesn't matter
for this handler but I thought I'd taste-test the pattern.

Signed-off-by: Joshua Marantz <jmarantz@google.com>