examples/front-proxy: base dockerfile on alpine
Previously, the example wouldn't build due to the wrong version of
python being used.

Rather than go through the pain of installing a modern version of python
on ubuntu:14.04, this patch just bases examples dockerfile on the envoy
alpine image which includes the necessary version of python3 by default.

Fixes #2262.

Signed-off-by: Louis Taylor <louis@kragniz.eu>