grpc: singleton manager/factory pattern for gRPC async clients.
As part of the switch to envoy::api::v2::GrpcService and preparing the code base for the Google C++
gRPC client, this PR introduces a singleton gRPC async client manager and factory pattern. This
allows for details of TLS and shared state across clients to be hidden from consumers and for
consumers to be agnostic about the type of gRPC client.

This PR also introduces the deprecation of the specification of gRPC services by cluster name,
envoy::api::v2::GrpcService should be used going forward.

Signed-off-by: Harvey Tuch <htuch@google.com>