proxy_protocol: Release file event when done.
Filter destruction is postponed, which makes it possible for the file
event of the proxy protocol filter to interfere with the file event of
the connection object that is created right after the proxy protocol
filter has successfully concluded. To fix this we explicitly destruct
the file event before returning from the proxy protocol filter.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>