proxy_protocol_test: Make tests more robust.
Some of the tests depended on the timing of data arraving to the
socket read by the proxy protocol filter. Make all tests run the
dispatcher in blocking mode, and stop the dispatcher when we know that
the proxy protocol processing has finished.

These changes allow these tests to pass on OSX, and also prevent
spurious test failures on other platforms.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>