Send optional entity-body with direct responses
*Description*
This change, which is part 2 of a 3-part implementation of the new direct
response feature, enables Envoy to include a body with a direct response.
The body can be specified inline within the configuration or loaded from
a file at request time.

*Risk Level*: Medium

*Testing*: Unit tests included

*Docs Changes*:
None yet. I'll make a docs PR once I've done the 3rd and last part of the
implementation: support for adding headers to direct responses.

*Release Notes*: N/A

*Issue*: #2315

*API Changes*: [#393](https://github.com/envoyproxy/data-plane-api/pull/393)

Signed-off-by: Brian Pane <bpane@pinterest.com>