bazel: fix protobuf UBSAN errors
*Description*:
The google-protobuf library, by default, uses unaligned reads on x86.
Those happen to work on that CPU architecture, but their behavior is
undefined in C/C++, so UBSAN (which runs along with ASAN in the CI
ASAN tests) treats them as an error.

Defining `ADDRESS_SANITIZER=1` causes the protobuf library to use
a portable alternative to the unaligned reads.

Signed-off-by: Brian Pane <bpane@pinterest.com>

*Risk Level*: Low

*Testing*: I tested this change with PR [2429](https://github.com/envoyproxy/envoy/pull/2429),
where it fixed the ASAN test failure.

*Docs Changes*: N/A

*Release Notes*: N/A