test: fix flake in access_log_integration_test.
Failed to update the stats check for this test in #2481.

In general, we should get better at abstracting away these gRPC client type dependent stats in
integration tests, but we should merge this PR to reduce CI flakes first.

Risk Level: Low
Testing: 100 runs of the test with -c dbg.

Signed-off-by: Harvey Tuch <htuch@google.com>