connection: Add support for socket options on created client connections (#2442)
Add an Options visitor to Network::ConnectionSocket. Listener filters
can set the visitor for any newly accepted connections, causing the
visitor is called in two sitations:

bool setOptions(ConnectionSocket&): This visitor method is called
right after a client socket is created for forwarding traffic received
from an accepted connection, before any traffic (including connection
setup) is sent on the new socket. This can be used to set socket
options, for example.

uint32_t hashKey(): This visitor is called on upstream connection pool
lookup. This gives the visitor the opportunity to separate upstream
connection pools so that connections for which a different values are
returned will not share connection pools. The purpose of this is to
allow the listener filter to make sure that a reused connection has
the required options as set by setOptions(). If this segmentation is
not needed the implementation should return zero.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>