Factor out most of the contents of main() into a new MainCommon class in main_common.cc.
This sets the stage for another PR which will use the factored out
MainCommon to performance test initialization without entering a main
loop.

Part of this PR is where we factor out all the initialization done
prior to running the server loop into MainCommon's ctor.  The listener
loop is started by a run() method on the class.

The orgaqnization is a little more complex than ideal because we want
to, for one round of import, support the existing calling pattern of
main_common, which leaves some of the initialization in main().

Another part of this CL is to fix a SEGV in the Server destruction
process, if you ever instantiate a Server without running the listener
loop and terminating via a signal or quitquitquit.

Signed-off-by: Joshua Marantz <jmarantz@google.com>