server: Two pass tag extraction (#2515)
Description:
This PR is one step toward resolving #2373, as outlined in a 2-page google doc (https://goo.gl/YxzBaB). It sets the stage for simplifying regexps by allowing them to all see the same input string, which is only mutated after all the tag extractions are done.

Risk Level: Medium. This is self-contained, easy to reason about, and well tested. But it is part of the startup flow and a failure stops envoy from serving traffic and deserves some scrutiny.

Testing: //test/... and speed-test of startup on 10k clusters.

Signed-off-by: Joshua Marantz <jmarantz@google.com>