hot-restarter.py: propagate SIGTERM to child processes
This allows hot-restarter.py to be used as a parent process to a
container engine (e.g. `runc`). Prior to this change, hot-restarter.py
would send a SIGKILL to all children when it receives a SIGTERM for
simplicity, because it assumed that its children were envoy processes
which have no state to clean up.

However, when running hot-restarter.py with children that have state
that should be cleaned up (e.g. container state), it's better to
propagate the SIGTERM in order to allow children to exit gracefully.

SIGTERM is now handled by propagating SIGTERM to all children, and then
waiting until either all children have exited gracefully or a constant
timeout has been hit, at which point the children are forcibly killed as
they were before.

Signed-off-by: Michael Puncel <mpuncel@squareup.com>