proxy_protocol: Mark the local address as restored if changed.
Change proxy protocol to set the local address if it really changes,
and mark the changed address as being restored. This allows proxy
protocol originated connections to be used with original destination
clusters. Original destination cluster only opens connections to
restored destination addresses, as otherwise the upstream connection
would connect back to the listener the downstream connection was
received at.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>